+++ packages/contracts/src/aelfContract.ts
@@ -1,12 +1,13 @@
-import { ChainId } from '@portkey/types';
-import { CallOptions, ContractProps, IPortkeyContract, SendOptions, SendResult, ViewResult } from './types';
-import { getTxResult, handleContractError, handleContractParams, handleFunctionName } from './utils';
 import { aelf, sleep } from '@portkey/utils';
+import { CallOptions, SendOptions, SendResult, ViewResult } from '@portkey/types';
+import { ChainId } from '@portkey/types';
+
+import { ContractProps, IPortkeyContract } from './types';
+import { getTxResult, handleContractError, handleContractParams, handleFunctionName } from './utils';
 import { BaseContract } from './baseContract';
 
 export class AElfContract extends BaseContract implements IPortkeyContract {
   public aelfContract: any;
-  public address: string;
   public aelfInstance?: any;
   public chainId?: ChainId;
   constructor(options: ContractProps) {