+++ packages/contracts/src/index.ts
@@ -1,6 +1,8 @@
-export * from './aelfContract';
-export * from './web3Contract';
-export * from './contract';
 export * from './utils';
+export * from './contract';
+export * from './caContract';
+export * from './web3Contract';
+export * from './aelfContract';
+export * from './baseContract';
 
 export * from './types';