+++ packages/contracts/src/types.ts
@@ -1,47 +1,5 @@
-import { ChainId, ChainType, IAElfRPCMethods } from '@portkey/types';
-
-export type SendOptions = {
-  from?: string;
-  gasPrice?: string;
-  gas?: number;
-  value?: number | string;
-  nonce?: number;
-  onMethod: 'transactionHash' | 'receipt' | 'confirmation';
-};
-
-export interface ErrorMsg {
-  name?: string;
-  code?: number;
-  message?: string;
-}
-export interface ViewResult<T = any> {
-  data?: T;
-  error?: ErrorMsg;
-}
-
-export interface SendResult<T = any> extends ViewResult<T> {
-  transactionId?: string;
-}
-
-export type ContractBasicErrorMsg = ErrorMsg;
-
-export type CallOptions = {
-  defaultBlock: number | string;
-  options?: any;
-  callback?: any;
-};
-export interface IContract {
-  address?: string;
-  chainId?: ChainId;
-  type: ChainType;
-  callViewMethod<T = any>(functionName: string, paramsOption?: any, callOptions?: CallOptions): Promise<ViewResult<T>>;
-  callSendMethod<T = any>(
-    functionName: string,
-    account: string,
-    paramsOption?: any,
-    sendOptions?: SendOptions,
-  ): Promise<SendResult<T>>;
-}
+import { ChainId, ChainType, IAElfRPCMethods, IBlockchainWallet, IContract, ViewResult } from '@portkey/types';
+import { IChain } from '@portkey/provider-types';
 
 export interface IPortkeyContract extends IContract {
   encodedTx<T = any>(functionName: string, paramsOption?: any): Promise<ViewResult<T>>;
@@ -56,5 +14,55 @@ export interface BaseContractOptions {
 
 export interface ContractProps extends BaseContractOptions {
   aelfContract?: any;
-  aelfInstance?: IAElfRPCMethods;
+  aelfInstance?: { chain: IAElfRPCMethods };
+}
+
+export interface CAContractProps extends ContractProps {
+  caContractAddress: string;
+  caContract: any;
+  caHash: string;
+}
+
+export type CallType = 'ca' | 'eoa';
+
+export interface IBaseOptions {
+  contractAddress: string;
+  callType?: CallType;
+  chianType?: ChainType;
+}
+
+export interface IProviderOptions extends IBaseOptions {
+  chainProvider: IChain;
+}
+
+export interface IEOAInstanceOptions extends IBaseOptions {
+  rpcUrl?: string;
+  aelfInstance?: { chain: IAElfRPCMethods };
+  account: { address: string } | IBlockchainWallet;
+}
+
+export interface ICAInstanceOptions extends IEOAInstanceOptions {
+  callType: 'ca';
+  caHash: string;
+  caContractAddress: string;
+}
+
+export interface IGetContract {
+  /**
+   * use provider contract
+   * @returns IContract
+   */
+  getContractBasic(options: IProviderOptions): Promise<IContract>;
+
+  /**
+   * use base contract
+   * @returns IPortkeyContract
+   */
+  getContractBasic(options: IEOAInstanceOptions): Promise<IPortkeyContract>;
+
+  /**
+   * use ca contract
+   * @returns IContract
+   */
+  getContractBasic(options: ICAInstanceOptions): Promise<IContract>;
 }