+++ CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.0-alpha.5](https://github.com/Portkey-Wallet/portkey-web/compare/v1.0.0-alpha.4...v1.0.0-alpha.5) (2023-06-30)
+
+### Features
+
+- 🎸 use ca contract ([6d75231](https://github.com/Portkey-Wallet/portkey-web/commit/6d75231a1ccae515b982be003a4a33b0462fc560))
+
+## [1.0.0-alpha.4](https://github.com/Portkey-Wallet/portkey-web/compare/v1.0.0-alpha.3...v1.0.0-alpha.4) (2023-06-29)
+
+**Note:** Version bump only for package root
+
+## [1.0.0-alpha.3](https://github.com/Portkey-Wallet/portkey-web/compare/v1.0.0-alpha.2...v1.0.0-alpha.3) (2023-06-29)
+
+### Bug Fixes
+
+- 🐛 export base contract ([beaa3a4](https://github.com/Portkey-Wallet/portkey-web/commit/beaa3a4f191e98d7f932b290442e712d42e016f9))
+
+## [1.0.0-alpha.2](https://github.com/Portkey-Wallet/portkey-web/compare/v1.0.0-alpha.1...v1.0.0-alpha.2) (2023-06-29)
+
+### Features
+
+- 🎸 did socket ([93e0d0a](https://github.com/Portkey-Wallet/portkey-web/commit/93e0d0a3d3a52cbafc33e1cf73ee5d31274e849f))
+- 🎸 socket ([5d456a0](https://github.com/Portkey-Wallet/portkey-web/commit/5d456a0aa157c5e307b55b9ae0e6fc10be337590))
+- 🎸 strict ([87c1182](https://github.com/Portkey-Wallet/portkey-web/commit/87c1182aebfd7f492bb9c93e2a0241da54f3c536))
+- 🎸 update dependencies ([4fc52d2](https://github.com/Portkey-Wallet/portkey-web/commit/4fc52d287ae65805f75ceaf4c5be31631d617f34))
+- 🎸 update readme ([cc24d77](https://github.com/Portkey-Wallet/portkey-web/commit/cc24d779371ee44fd3e73f16c9bb3ae37368393b))
+
+### Bug Fixes
+
+- 🐛 host ([8252fda](https://github.com/Portkey-Wallet/portkey-web/commit/8252fda775f9e0a11a50184c6a6c08be9a17b822))
+- 🐛 test ([9be6f40](https://github.com/Portkey-Wallet/portkey-web/commit/9be6f40571361a092ec2e444b9cb03d84fed719a))
+- 🐛 test ([c907dd1](https://github.com/Portkey-Wallet/portkey-web/commit/c907dd1f2b00d1a6b41ef315c5a6c203f40af79f))
+
 ## [1.0.0-alpha.1](https://github.com/Portkey-Wallet/portkey-web/compare/v1.0.0-alpha.0...v1.0.0-alpha.1) (2023-06-28)
 
 ### Bug Fixes